refactor to split out main method
